job_agent/requirements.txt
# 核心依赖
playwright==1.40.0
anthropic==0.18.0
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0

# 数据处理
pandas==2.1.4
sqlalchemy==2.0.25

# 工具库
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2

# 可选:PDF生成
markdown==3.5.1
